Nested loops
Nested Loops:-A nested loop refers to a loop inside another loop. In other words, it is when one loop is placed within the body of another loop. The outer loop runs a set number of times, and for each iteration of the outer loop, the inner loop executes its set of instructions.
Application:-Matrix operations , Pattern printing , Multi Dimensional Iteration...